+++ test-core/matrix/main_test_matrix.cpp
@@ -409,7 +409,7 @@ TEST(Iterative, MinresQLP)
                 columns.push_back(i + j);
                 values.push_back(val[j + 2]);
             }
-        rowPointers[i + 1] = values.size();
+        rowPointers[i + 1] = (int)values.size();
     }
     corecvs::SparseMatrix M(N, N, values, columns, rowPointers);
     corecvs::Vector xx(N);
@@ -436,7 +436,7 @@ TEST(Iterative, MinresQLPPreconditioned)
                 columns.push_back(i + j);
                 values.push_back(val[j + 2]);
             }
-        rowPointers[i + 1] = values.size();
+        rowPointers[i + 1] = (int)values.size();
     }
     corecvs::SparseMatrix M(N, N, values, columns, rowPointers);
     corecvs::Vector xx(N);
